requirements.txt
spacy
numpy
pandas
sentence-transformers
streamlit
faiss-cpu
nltk
matplotlib
rapidfuzz
openai
langchain
python-dotenv
langchain-community
